Add partial finished heater-application
