+++ communication/esp8266/arduino/devices/boilerplates/B-103-oleddisplay/platformio.ini
@@ -12,8 +12,6 @@ build_flags = -D PB_ENABLE_MALLOC
 upload_port = /dev/ttyUSB0
 lib_extra_dirs = ../../../libraries/
 extra_script = ../../../../scripts/proto_generator.py
-custom_option = linux
-
 
 [env:esp12e]
 platform = espressif8266
@@ -25,7 +23,6 @@ upload_port = ${common.upload_port}
 lib_extra_dirs = ${common.lib_extra_dirs}
 lib_deps = ${common.lib_deps}
 extra_script = ${common.extra_script}
-custom_option = ${common.custom_option}
 
 [env:esp01]
 platform = espressif8266
@@ -37,4 +34,3 @@ upload_port = ${common.upload_port}
 lib_extra_dirs = ${common.lib_extra_dirs}
 lib_deps = ${common.lib_deps}
 extra_script = ${common.extra_script}
-custom_option = ${common.custom_option}
