+++ communication/generic/python/entities/heater-control/.gitignore
@@ -1,4 +1,3 @@
 #Generated proto files
 messages_pb2.py
-proto_nab.py
 configuration.json
