+++ communication/generic/python/entities/heater-control/configuration_default.json
@@ -1,5 +1,4 @@
 {
-    "os": "linux",
     "mqtt" : {
         "clientId" : "heatercontrol",
         "broker" : "thisIsYourBrokerDomainAddressOrIp",
