+++ communication/generic/python/entities/heater-control/heatercontrol/__init__.py
@@ -1,4 +1,7 @@
 import ioant.proto.proto as proto
+import sys
 
 proto.python_main("../../../../../common/proto/",
                   "heatercontrol/generated_proto/.")
+
+sys.path.append("heatercontrol/generated_proto/")
