+++ communication/generic/python/entities/heater-control/heatercontrol/generated_proto/.gitignore
@@ -1,0 +1,4 @@
+*
+*/
+!__init__.py
+!.gitignore
