+++ communication/generic/python/entities/heater-control/logs/.gitignore
@@ -1,0 +1,3 @@
+*
+*/
+!.gitignore
