+++ communication/generic/python/entities/heater-control/main.py
@@ -2,7 +2,14 @@ import heatercontrol.heatercontrol as control
 import sys
 from ioant.utils import utils
 import os
+import logging
 
+logging.basicConfig(filename='logs/output.log',
+                    level=logging.DEBUG,
+                    format='%(asctime)s %(name)-5s %(levelname)-8s %(message)s',
+                    datefmt='%m-%d %H:%M')
+console = logging.StreamHandler()
+logging.getLogger('').addHandler(console)
 
 def init_ascii():
     message = "\
@@ -11,26 +18,16 @@ def init_ascii():
 ========================================================================="
     return message
 
+
 relative_path_steps = "../../../../../"
 
 if __name__ == "__main__":
     print(init_ascii())
-
+    logging.info('Starting application')
     script_dir = os.path.dirname(os.path.realpath(__file__))
-    db_schema_path = utils.return_absolut_path(script_dir, relative_path_steps
-                                               + "storage/schema.json")
-    db_schema_dict = utils.fetch_json_file_as_dict(db_schema_path)
     configuration_path = utils.return_absolut_path(os.path.dirname(__file__),
                                                    'configuration.json')
-    configuration_dict = utils.fetch_json_file_as_dict(configuration_path)
-
-    mqtt_broker = configuration_dict['mqtt']['broker']
-    mqtt_port = configuration_dict['mqtt']['port']
-
-    client = control.initiate_client(mqtt_broker,
-                                       mqtt_port,
-                                       configuration_dict)
-    if client:
-        control.loop_mqtt_client(client)
-
+    configuration = utils.fetch_json_file_as_dict(configuration_path)
+    control.setup(configuration)
+    control.loop()
     sys.exit()
