+++ communication/raspberry_pi/python/devices/boilerplates/lenscam/lenscam/__init__.py
@@ -1,4 +1,5 @@
 import ioant.proto.proto as proto
-
+import sys
 proto.python_main("../../../../../../common/proto/",
                   "lenscam/generated_proto/.")
+sys.path.append("heatercontrol/generated_proto/")
