+++ communication/raspberry_pi/python/devices/boilerplates/lenscam/logs/.gitignore
@@ -1,0 +1,3 @@
+*
+*/
+!.gitignore
