+++ communication/raspberry_pi/python/devices/boilerplates/lenscam/main.py
@@ -2,6 +2,15 @@ import lenscam.lenscam as lenscam
 import sys
 from ioant.utils import utils
 import os
+import logging
+
+logging.basicConfig(filename='logs/output.log',
+                    level=logging.DEBUG,
+                    format='%(asctime)s %(name)-5s %(levelname)-8s %(message)s',
+                    datefmt='%m-%d %H:%M')
+console = logging.StreamHandler()
+logging.getLogger('').addHandler(console)
+
 
 def init_ascii():
     message = "\
@@ -14,18 +23,11 @@ def init_ascii():
 if __name__ == "__main__":
     print(init_ascii())
 
+    logging.info('Starting application')
     script_dir = os.path.dirname(os.path.realpath(__file__))
     configuration_path = utils.return_absolut_path(os.path.dirname(__file__),
                                                    'configuration.json')
-    configuration_dict = utils.fetch_json_file_as_dict(configuration_path)
-
-    mqtt_broker = configuration_dict['mqtt']['broker']
-    mqtt_port = configuration_dict['mqtt']['port']
-
-    client = lenscam.initiate_client(mqtt_broker,
-                                      mqtt_port,
-                                      configuration_dict)
-    if client:
-        lenscam.loop_mqtt_client(client)
-
+    configuration = utils.fetch_json_file_as_dict(configuration_path)
+    lenscam.setup(configuration)
+    lenscam.loop()
     sys.exit()
